Fix so default config do not overwrite env config
The default config was added as a fallback in the config from
`package.json` and since that config takes precedence over config
from environmental variables the default config was always used
instead of the environmental variables.

This moves the defaults values setup so they are only used of no
other config is provided for the value.